feat: 이력서 수정 및 삭제 기능 구현

+++ src/main/java/com/example/study_mate/resume/repository/ResumeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ResumeRepository extends JpaRepository<Resume, Integer> {
+public interface ResumeRepository extends JpaRepository<Resume, Long> {
 }
